Call wallet service on transfer ownership
