+++ proto/wallet_service.proto
@@ -35,6 +35,10 @@ message ActivateWalletRequest {
     string hash = 3;
 }
 
+message CoopRequest {
+    string coop = 1;
+}
+
 message Empty {
 }
 
@@ -42,4 +46,5 @@ service WalletService {
     rpc GetWalletsByOwner (GetWalletsByOwnerRequest) returns (WalletsResponse);
     rpc GetWalletsByHash (GetWalletsByHashRequest) returns (WalletsResponse);
     rpc ActivateWallet (ActivateWalletRequest) returns (Empty);
+    rpc UpdateCoopRoles (CoopRequest) returns (Empty);
 }
