+++ test/util/walletSvcGrpc.js
@@ -1,7 +1,6 @@
 const { createMockServer } = require("grpc-mock")
 const path = require('path')
 const config = require('../../config')
-const globalSetup = require('../global-setup')
 
 let walletGrpcMockServer
 
@@ -12,7 +11,8 @@ async function init() {
         packageName: 'com.ampnet.walletservice.proto',
         serviceName: 'WalletService',
         rules: [
-            { method: "activateWallet", input: ".*", output: { } }
+            { method: "activateWallet", input: ".*", output: { } },
+            { method: "updateCoopRoles", input: ".*", output: { } }
         ]
     })
     walletGrpcMockServer.listen(config.get().walletServiceGrpc)
